fix(web): use consistent navbar across Articles and Medical Glossary pages
